Log file download for Repository log files do not work from Share. This is due to the filePath in the request not being resolved to an absolute path (no / in start of string). With this it fails to resolve as a valid path to download from.
This is resolved by checking if it starts with /, if not add it.
Moved this into the AbstractLogFileWebScript to not duplicate code.

Could probably be resolved by fixing the substring, but it can be that
different browsers handles double slashes (that is actually in the
request from Share) differently. So using this method instead.